Изменены модули

+++ backendApi/src/modules/priority/priority.controller.ts
@@ -1,9 +1,9 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req, HttpException, HttpStatus } from '@nestjs/common';
 import { PriorityService } from './priority.service';
-import { CreatePriorityDto } from './dto/create-priority.dto';
-import { UpdatePriorityDto } from './dto/update-priority.dto';
+import { CreatePriorityDto, UpdatePriorityDto } from './dto';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { JwtAuthGuard } from 'src/modules/auth/guards/auth.guard';
+import { AppError } from 'src/common/constants/error';
 
 @ApiBearerAuth()
 @Controller('priority')
@@ -24,13 +24,25 @@ export class PriorityController {
 
   @UseGuards(JwtAuthGuard)
   @Patch()
-  update(@Body() updatePriorityDto: UpdatePriorityDto, @Req() request) {
+  async update(@Body() updatePriorityDto: UpdatePriorityDto, @Req() request) {
+    let foundPriority = null;
+    if (updatePriorityDto.priority_id) {
+      foundPriority = await this.priorityService.findOne(updatePriorityDto.priority_id);
+    }
+    if (!foundPriority) {
+      throw new HttpException(AppError.PRIORITY_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
     return this.priorityService.update(updatePriorityDto, request.user.user_id);
   }
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string, @Req() request) {
+  async remove(@Param('id') id: number, @Req() request) {
+    const foundPriority = await this.priorityService.findOne(id);
+    if (foundPriority == null) {
+      throw new HttpException(AppError.PRIORITY_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
     return this.priorityService.remove(+id, request.user.user_id);
   }
 }
